requirements.txt
pydantic-settings==2.12.0
djangorestframework==3.16.1
google-genai==1.52.0
drf-spectacular==0.29.0
pydantic==2.12.5